+++ src/components/organisms/CommentList/index.scss
@@ -1,0 +1,5 @@
+.comment--list {
+  display: flex;
+  flex-direction: column;
+  gap: 15px;
+}
