+++ src/components/organisms/CommentList/index.tsx
@@ -1,0 +1,3 @@
+'use client'
+
+export { default as CommentList } from './CommentList'
